+++ src/Custom/Assistants/AssistantClient.cs
@@ -903,6 +903,7 @@ public partial class AssistantClient
             runOptions.MaxCompletionTokens,
             runOptions.TruncationStrategy,
             runOptions.ToolConstraint,
+            runOptions.ParallelToolCallsEnabled,
             runOptions.ResponseFormat,
             serializedAdditionalRawData: null);
         return internalRequest.ToBinaryContent();