+++ src/Custom/VectorStores/VectorStoreFileAssociation.cs
@@ -23,4 +23,7 @@ public partial class VectorStoreFileAssociation
     /// </summary>
     [CodeGenMember("UsageBytes")]
     public int Size { get; }
+
+    [CodeGenMember("ChunkingStrategy")]
+    public FileChunkingStrategy ChunkingStrategy { get; }
 }